requirements.txt
# Shared
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
python-dotenv>=1.0,<2.0

# Memory Agent
chromadb-client>=0.5,<1.0
neo4j>=5.0,<6.0
openai>=1.0,<2.0

# Scraper Agent
httpx>=0.27,<1.0
newspaper3k>=0.2.8,<0.3
praw>=7.0,<8.0
feedparser>=6.0,<7.0
apscheduler>=3.10,<4.0
lxml_html_clean>=0.1.0

# Preprocessing Agent
spacy>=3.7,<4.0

# Dev / Testing
pytest>=8.0,<9.0
pytest-cov>=5.0,<6.0
respx>=0.21,<1.0
ruff>=0.4,<1.0
